requirements.txt
openai>=1.14.0
hubspot-api-client>=9.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.32.0
python-dotenv>=1.0.1
pydantic>=2.6.0
httpx>=0.27.0
aiosqlite>=0.20.0
plotly>=5.20.0
